Some changes at LE extension points architecture.
Now, the extention must have the editor-ref attribute to define de extension target. If no reference is passed, all LuaEditor instances will receive the extension.

With this change, extend the LuaEditor is not completely necessary. Itś just need to declar another editor with a diferent Id and contribute with him direct.